chore: refactor pkg/client/oci for future oci-sif work
